+++ lib/dasl/car/encoder.ex
@@ -19,17 +19,17 @@ defmodule DASL.CAR.Encoder do
   def encode(%CAR{version: 1, roots: roots, blocks: blocks}, opts \\ []) do
     verify = Keyword.get(opts, :verify, true)
 
-    with {:ok, header_bin} <- encode_header(roots),
-         {:ok, blocks_bin} <- encode_blocks(blocks, verify) do
-      {:ok, header_bin <> blocks_bin}
+    with {:ok, header_iodata} <- encode_header(roots),
+         {:ok, blocks_iodata} <- encode_blocks(blocks, verify) do
+      {:ok, IO.iodata_to_binary([header_iodata, blocks_iodata])}
     end
   end
 
-  @spec encode_header(list(CID.t())) :: {:ok, binary()} | header_error()
+  @spec encode_header(list(CID.t())) :: {:ok, iodata()} | header_error()
   defp encode_header(roots) do
     with :ok <- validate_roots(roots),
          {:ok, metadata_bin} <- DRISL.encode(%{"version" => 1, "roots" => roots}) do
-      {:ok, LEB128.encode(byte_size(metadata_bin)) <> metadata_bin}
+      {:ok, [LEB128.encode(byte_size(metadata_bin)), metadata_bin]}
     else
       {:error, :header, _} = err -> err
       {:error, reason} -> {:error, :header, reason}
@@ -47,20 +47,20 @@ defmodule DASL.CAR.Encoder do
     end
   end
 
-  @spec encode_blocks(map(), boolean()) :: {:ok, binary()} | block_error()
+  @spec encode_blocks(map(), boolean()) :: {:ok, iodata()} | block_error()
   defp encode_blocks(blocks, verify) do
-    Enum.reduce_while(blocks, {:ok, <<>>}, fn {cid, data}, {:ok, acc} ->
+    Enum.reduce_while(blocks, {:ok, []}, fn {cid, data}, {:ok, acc} ->
       case encode_block(cid, data, verify) do
-        {:ok, block_bin} -> {:cont, {:ok, acc <> block_bin}}
+        {:ok, block_iodata} -> {:cont, {:ok, [acc | [block_iodata]]}}
         {:error, _, _} = err -> {:halt, err}
       end
     end)
   end
 
-  @spec encode_block(CID.t(), binary(), boolean()) :: {:ok, binary()} | block_error()
+  @spec encode_block(CID.t(), binary(), boolean()) :: {:ok, iodata()} | block_error()
   defp encode_block(%CID{bytes: cid_bytes} = cid, data, verify) when is_binary(data) do
     with :ok <- maybe_verify(verify, cid, data) do
-      {:ok, LEB128.encode(byte_size(cid_bytes) + byte_size(data)) <> cid_bytes <> data}
+      {:ok, [LEB128.encode(byte_size(cid_bytes) + byte_size(data)), cid_bytes, data]}
     end
   end